+++ OndselSolver/PartFrame.cpp
@@ -150,7 +150,7 @@ void PartFrame::removeRedundantConstraints(std::shared_ptr<std::vector<int>> red
 		redunCon->constraint = aGeu;
 		aGeu = redunCon;
 	}
-	for (size_t i = 0; i < aGabs->size(); i++)
+	for (int i = 0; i < aGabs->size(); i++)
 	{
 		auto& constraint = aGabs->at(i);
 		if (std::find(redundantEqnNos->begin(), redundantEqnNos->end(), constraint->iG) != redundantEqnNos->end()) {
@@ -164,7 +164,7 @@ void PartFrame::removeRedundantConstraints(std::shared_ptr<std::vector<int>> red
 void PartFrame::reactivateRedundantConstraints()
 {
 	if (aGeu->isRedundant()) aGeu = std::dynamic_pointer_cast<RedundantConstraint>(aGeu)->constraint;
-	for (size_t i = 0; i < aGabs->size(); i++)
+	for (int i = 0; i < aGabs->size(); i++)
 	{
 		auto& con = aGabs->at(i);
 		if (con->isRedundant()) {
